refactor(store-core): extract cost/usage handlers into handlers/usage.ts (audit P2-3)
Ninth slice of the P2-3 split. Move the contiguous cost/usage message
handlers (cost_update / session_usage) into a new handlers/usage.ts.

Pure move, no logic change: the moved section is byte-identical to the
origin/main block (proven by diff). Imports CumulativeUsage (../types) and
SessionPatch (./_shared); both remain used elsewhere in index.ts. The
barrel re-exports via `export * from './usage'`, so the public surface is
unchanged.

handleResultUsage and handleSessionCostThresholdCrossed are not contiguous
(they sit in the result/stream area) and stay in index.ts for now.

handlers/index.ts: 4695 -> 4616 lines.
